Make the URL lookup mechanism into a reusable URLMixin
