+++ common/models/__init__.py
@@ -2,3 +2,4 @@ from .location import LocationMixin  # noqa
 from .lockable import Lockable  # noqa
 from .prefetch_snooping import PrefetchSnoopingMixin  # noqa
 from .thumbnails import ModelWithThumbnails  # noqa
+from .urls import URLMixin  # noqa
